+++ mobile/src/androidTest/java/org/openhab/habdroid/TestWithoutIntro.java
@@ -1,0 +1,30 @@
+package org.openhab.habdroid;
+
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
+import android.support.test.InstrumentationRegistry;
+
+import org.junit.Before;
+import org.openhab.habdroid.util.Constants;
+
+public abstract class TestWithoutIntro extends ProgressbarAwareTest {
+    @Before
+    public void setup() {
+        SharedPreferences.Editor edit = PreferenceManager
+                .getDefaultSharedPreferences(InstrumentationRegistry.getTargetContext())
+                .edit();
+
+        edit.putString(Constants.PREFERENCE_SITEMAP, "");
+        if (preselectSitemap()) {
+            edit.putString(Constants.PREFERENCE_SITEMAP, "demo");
+        }
+
+        edit.putBoolean(Constants.PREFERENCE_FIRST_START, false).commit();
+
+        super.setup();
+    }
+
+    protected boolean preselectSitemap() {
+        return false;
+    }
+}
