+++ mobile/src/androidTest/java/org/openhab/habdroid/ui/BasicWidgetTest.java
@@ -1,37 +1,26 @@
 package org.openhab.habdroid.ui;
 
 
-import android.preference.PreferenceManager;
 import android.support.constraint.ConstraintLayout;
-import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.DataInteraction;
-import android.support.test.espresso.IdlingResource;
 import android.support.test.espresso.ViewInteraction;
-import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
 import android.view.View;
 
 import org.hamcrest.Matchers;
 import org.hamcrest.core.IsInstanceOf;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.openhab.habdroid.OpenHABProgressbarIdlingResource;
+import org.openhab.habdroid.TestWithoutIntro;
 import org.openhab.habdroid.R;
-import org.openhab.habdroid.util.Constants;
 
 import static android.support.test.espresso.Espresso.onData;
 import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.Espresso.registerIdlingResources;
-import static android.support.test.espresso.Espresso.unregisterIdlingResources;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
-import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withParent;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
@@ -42,33 +31,10 @@ import static org.openhab.habdroid.TestUtils.childAtPosition;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class BasicWidgetTest {
-
-    @Rule
-    public ActivityTestRule<OpenHABMainActivity> mActivityTestRule = new ActivityTestRule<>
-            (OpenHABMainActivity.class, true, false);
-
-    private IdlingResource mProgressbarIdlingResource;
-
-    @Before
-    public void setup() {
-        PreferenceManager
-                .getDefaultSharedPreferences(InstrumentationRegistry.getTargetContext())
-                .edit()
-                .putString(Constants.PREFERENCE_SITEMAP, "")
-                .putBoolean(Constants.PREFERENCE_FIRST_START, false)
-                .commit();
-
-        mActivityTestRule.launchActivity(null);
-    }
+public class BasicWidgetTest extends TestWithoutIntro {
 
     @Test
     public void openHABMainActivityTest() throws InterruptedException {
-        View progressBar = mActivityTestRule.getActivity().findViewById(R.id.toolbar_progress_bar);
-        mProgressbarIdlingResource = new OpenHABProgressbarIdlingResource("Progressbar " +
-                "IdleResource", progressBar);
-        registerIdlingResources(mProgressbarIdlingResource);
-
         // do we have sitemap selection popup?
         ViewInteraction linearLayout = onView(
                 Matchers.allOf(IsInstanceOf.<View>instanceOf(android.widget.LinearLayout.class),
@@ -107,10 +73,6 @@ public class BasicWidgetTest {
                                 1),
                         isDisplayed()));
         mainmenu.check(matches(withText("Main Menu")));
-
-        ViewInteraction voice = onView(
-                allOf(withId(R.id.mainmenu_voice_recognition), withContentDescription("Voice recognition"), isDisplayed()));
-        voice.check(matches(isDisplayed()));
 
         ViewInteraction firstfloor = onView(
                 allOf(withId(R.id.widgetlabel), withText("First Floor"),
@@ -251,11 +213,5 @@ public class BasicWidgetTest {
         ViewInteraction imageButton2 = onView(
                 Matchers.allOf(withId(R.id.rollershutterbutton_stop), isDisplayed()));
         imageButton2.check(matches(isDisplayed()));
-    }
-
-    @After
-    public void unregisterIdlingResource() {
-        if (mProgressbarIdlingResource != null)
-            unregisterIdlingResources(mProgressbarIdlingResource);
     }
 }
